rename LabelCreator to avoid collision with legacy version of this library
